refactor: Rename for consistency
