+++ src/app/Layout/Presentation/Views/index.ts
@@ -2,4 +2,4 @@ export * from "./DistrictOverview";
 export * from "./ElectionOverview";
 export * from "./SeatDistribution";
 export * from "./SeatsPerParty";
-export * from "./SingleCounty";
+export * from "./SingleDistrict";
